+++ v-next/example-project/contracts/UserRemappedImport.sol
@@ -2,3 +2,5 @@
 pragma solidity ^0.8.0;
 
 import "remapped/Ownable.sol";
+
+library UserRemappedImport {}