+++ v-next/hardhat-verify/test/utils.ts
@@ -3,6 +3,7 @@ import type {
   Interceptable,
   TestDispatcher,
 } from "@nomicfoundation/hardhat-utils/request";
+import type { ResolvedConfigurationVariable } from "hardhat/types/config";
 import type { HardhatRuntimeEnvironment } from "hardhat/types/hre";
 import type {
   EthereumProvider,
@@ -176,5 +177,34 @@ export async function deployContract(
       return (receipt as any).contractAddress;
     }
     await new Promise((r) => setTimeout(r, 1000));
+  }
+}
+
+export class MockResolvedConfigurationVariable
+  implements ResolvedConfigurationVariable
+{
+  public _type: "ResolvedConfigurationVariable" =
+    "ResolvedConfigurationVariable";
+  public format: string = "{variable}";
+  readonly #value: string;
+
+  constructor(value: string) {
+    this.#value = value;
+  }
+
+  public async get(): Promise<string> {
+    return this.#value;
+  }
+
+  public async getUrl(): Promise<string> {
+    return this.#value;
+  }
+
+  public async getBigInt(): Promise<bigint> {
+    return BigInt(this.#value);
+  }
+
+  public async getHexString(): Promise<string> {
+    return this.#value;
   }
 }