+++ v-next/hardhat-viem-assertions/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nomicfoundation/hardhat-viem-assertions
 
+## 3.0.4
+
+### Patch Changes
+
+- 67008ab: Improved the documentation
+
 ## 3.0.3
 
 ### Patch Changes