+++ v-next/hardhat/CHANGELOG.md
@@ -1,5 +1,35 @@
 # hardhat
 
+## 3.0.15
+
+### Patch Changes
+
+- 9fb054a: Fix the initialization of the Mocha and Ethers sample project when using pnpm
+
+## 3.0.14
+
+### Patch Changes
+
+- 3d42a6f: Remove the compilation spinner when the compilation fails
+- 52e1eb4: Fix the `test solidity` task's compilation process
+- ce5c22a: Fail when a file isn't built neither as contract nor test
+
+## 3.0.13
+
+### Patch Changes
+
+- 6e4ad6d: Fixed unexpected test failure when running in isolate/gas stats mode ([#1146](https://github.com/NomicFoundation/edr/issues/1146))
+- 6e4ad6d: Added latest dynamic base fee parameters to Base Mainnet chain config ([#1141](https://github.com/NomicFoundation/edr/pull/1141))
+
+## 3.0.12
+
+### Patch Changes
+
+- 147f8a6: Improved the look and feel of compilation output ([#7669](https://github.com/NomicFoundation/hardhat/pull/7669))
+- 50ce3e0: Fix a bug that could lead to different builds having the same build id. Note that this version will change how every build id is computed.
+- 31f0b03: Don't compile solidity tests files when not needed ([#7669](https://github.com/NomicFoundation/hardhat/pull/7669))
+- a77c06a: Added a new minimal sample project
+
 ## 3.0.11
 
 ### Patch Changes